+++ src/server/config.ts
@@ -55,6 +55,7 @@ export type QConfig = {
     }
     remp: RempConfig
     useListeners?: boolean
+    walkingUseCache: boolean
     subscriptionsMode: SubscriptionsMode
     blockchain: QBlockchainDataConfig
     counterparties: string[]
@@ -88,6 +89,7 @@ export type QConfig = {
 
 export type FilterConfig = {
     orConversion: FilterOrConversion
+    stringifyKeyInAqlComparison: boolean
 }
 
 export type QBlockchainDataConfig = {
@@ -230,6 +232,13 @@ export const configParams = {
                     " and combines results (this option provides faster execution\n" +
                     " than OR operator in AQL)",
             ),
+            stringifyKeyInAqlComparison: ConfigParam.boolean(
+                "stringify-key-in-aql-comparison",
+                false,
+                "**UNSTABLE!** If `true` then AQL will use `TO_STRING(doc._key)` conversion\n" +
+                    'if _key comparison operator is used in filter (e.g. `{ id: { lt: "123" }`).',
+                false,
+            ),
         },
         maxRuntimeInS: ConfigParam.integer(
             "query-max-runtime",
@@ -257,6 +266,11 @@ export const configParams = {
         true,
         "Use database listeners for subscriptions (deprecated in favor of subscriptions-mode)",
         true,
+    ),
+    walkingUseCache: ConfigParam.boolean(
+        "walking-use-cache",
+        false,
+        "Use cache to serve block walking algorithm",
     ),
     ignoreMessagesForLatency: ConfigParam.boolean(
         "ignore-messages-for-latency",