+++ src/server/filter/filters.ts
@@ -89,6 +89,7 @@ export class QExplanation {
 
 export type QParamsOptions = {
     explain?: boolean
+    stringifyKeyInAqlComparison?: boolean
 }
 
 /**
@@ -98,12 +99,15 @@ export class QParams {
     values: Record<string, unknown>
     count: number
     explanation: QExplanation | null
+    stringifyKeyInAqlComparison: boolean
 
     constructor(options?: QParamsOptions) {
         this.count = 0
         this.values = {}
         this.explanation =
             options && options.explain ? new QExplanation() : null
+        this.stringifyKeyInAqlComparison =
+            options?.stringifyKeyInAqlComparison ?? false
     }
 
     clear() {
@@ -314,6 +318,7 @@ function filterConditionOp(
      */
 
     const isKeyOrderedComparison =
+        params.stringifyKeyInAqlComparison &&
         (path === "_key" || path.endsWith("._key")) &&
         op !== "==" &&
         op !== "!="